fix nanobind migration issues, drop unused pybind11 dep
- collect_indices only accepted python int/iterable, silently rejecting
  numpy integer scalars (e.g. np.int64) that worked under pybind11; use
  nb::try_cast so numpy scalars round-trip like before
- remove two copy-paste duplicate lines left over from the pybind11 ->
  nanobind port in qis()/qisi()
- drop the now-unused pybind11 dev dependency
- update doc/type-stubs.md for the nanobind stubgen workflow (README
  already documented it, this doc didn't)
- add 2.5.0 release notes for the build system migration

+++ tests/test_all.py
@@ -175,8 +175,8 @@ def test_IPF() -> None:
     assert np.allclose(np.sum(p, 0), m1)
     assert np.allclose(np.sum(p, 1), m0)
 
-    # mixed type indices
-    im = ((0,), np.array([1]), 2)
+    # mixed type indices, including a numpy integer scalar
+    im = ((0,), np.array([1]), np.int64(2))
     s = np.array([[[1.0, 1.0], [1.0, 1.0]], [[1.0, 1.0], [1.0, 1.0]]])
     p, stats = hl.ipf(s, im, (m0, m1, m2))  # ty:ignore[invalid-argument-type]
     assert stats["conv"]
